Serch color prob

+++ color/color.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   color.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: rmatsuba <rmatsuba@student.42tokyo.jp>     +#+  +:+       +#+        */
+/*   By: yoshidakazushi <yoshidakazushi@student.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/24 11:23:25 by rmatsuba          #+#    #+#             */
-/*   Updated: 2024/08/10 18:19:42 by rmatsuba         ###   ########.fr       */
+/*   Updated: 2024/09/27 11:50:43 by yoshidakazu      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -40,20 +40,19 @@ bool	check_color(int color)
 	return (true);
 }
 
-unsigned int	get_color(char *str)
+double	get_color(char *str)
 {
 	unsigned int	color;
 	
 	color = ft_atoi(str);
 	if (!check_color(color))
 		return (300);
-	return (color);
+	return ((double)color);
 }
 
 void	set_rgb(char *str, t_rgb *rgb)
 {
 	char	**str_rgb;
-
 	str_rgb = ft_split(str, ',');
 	rgb->r = get_color(str_rgb[0]);
 	rgb->g = get_color(str_rgb[1]);
